fix(CombatStats): correct overkill damage calculation on client
Previously, overkill damage was incorrectly calculated on clients because
the health extension was not yet synchronized when add_attack_result was
called via RPC. This caused the overkill value to be underestimated.

This change adjusts the calculation by adding the current damage to the
overkill value on clients, ensuring accurate tracking. On the server,
the calculation remains unchanged as damage is already applied before
add_attack_result is called.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>